config: fix SSL renewal
We're not using `certbot renew`, so a new folder gets created with
a numeric suffix on each iteration. Blow away the old folder and
discover the suffix of the new one using globbing.

This is a hack, but probably good enough (and permits us not to
distinguish between initial run and renewal, which is convenient).